Terraform dev env working with M1 mac
The Github Terraform module does not support darwin_arm.
This makes it hard to work with the terraform code locally.

This PR gets around that by creating a docker image that
one can run Terraform through if need be.

The ergonomics are decent, so it's an OK workaround
until it starts working natively again.

This is a prerequisite for #192.