Migrations only run once, on service startup
The old `singleton`-pattern only worked for calls within the same file.
Migrations are now moved one abstraction layer above, and are called on
server startup, separate from other database logic.

The logic is coupled with functions for booting a server, in a new
`bootWithMigrations`-function. I'm not sure how I feel about that, but
I at least think it's an improvement.

There's still very many ways to make migrations safer in this system,
but the current one is OK for now. Thus this closes #194.